Don't error on tag pushes
Previously we would run `abort()` if the ref didn't refer to a head.

Our Jenkins instance pushes tags to GitHub which triggers this webhook
again. Instead of doing a 400 Bad Request for those we should just silently
ignore them. This makes the GitHub webhook interface green and stops it
looking like our webhook is erroring.